fix: rename cmd directory from teltonika_go to teltonika-go for consistency
